# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to provide food service by users\n\n\nThis should show the struct of endpoints",
    "title": "Food API.",
    "contact": {
      "url": "https://www.linkedin.com/in/daniel-de-la-pava-suarez/"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/api/v1",
  "paths": {
...
    "/users": {
      "get": {
        "description": "Returns all user in the database",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "User"
        ],
        "summary": "GetAllUserHandler.",
        "operationId": "getAllUser",
        "parameters": [
          {
            "type": "string",
            "description": "type: apiKey",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SwaggerAllUserResponse"
          },
          "401": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "404": {
            "$ref": "#/responses/SwaggerErrorMessage"
          }
        }
      },
      "post": {
        "description": "Create a new user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "User"
        ],
        "summary": "CreateHandler.",
        "operationId": "userRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/SwaggerUserResponse"
          },
          "400": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "409": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "422": {
            "$ref": "#/responses/SwaggerErrorMessage"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Response one user by id",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "User"
        ],
        "summary": "GetOneHandler.",
        "operationId": "idUserPath",
        "parameters": [
          {
            "type": "string",
            "description": "type: apiKey",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SwaggerUserResponse"
          },
          "400": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "401": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "404": {
            "$ref": "#/responses/SwaggerErrorMessage"
          }
        }
      },
      "put": {
        "description": "Update a stored user by id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "User"
        ],
        "summary": "UpdateHandler.",
        "operationId": "userUpdateRequest",
        "parameters": [
          {
            "type": "string",
            "description": "type: apiKey",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "names",
                "last_names",
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                },
                "last_names": {
                  "type": "string",
                  "x-go-name": "LastNames"
                },
                "names": {
                  "type": "string",
                  "x-go-name": "Names"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SwaggerUserResponse"
          },
          "400": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "401": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "409": {
            "$ref": "#/responses/SwaggerErrorMessage"
          },
          "422": {
            "$ref": "#/responses/SwaggerErrorMessage"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ErrorMessage": {
      "description": "ErrorMessage structure that returns errors",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "food-api/infrastructure/middleware"
    },
...
    "User": {
      "description": "Data of User",
      "type": "object",
      "required": [
        "names",
        "last_names",
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "last_names": {
          "type": "string",
          "x-go-name": "LastNames"
        },
        "names": {
          "type": "string",
          "x-go-name": "Names"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "food-api/domain/user/domain/model"
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "last_names": {
          "type": "string",
          "x-go-name": "LastNames"
        },
        "names": {
          "type": "string",
          "x-go-name": "Names"
        }
      },
      "x-go-package": "food-api/domain/user/application/v1/response"
    }
  },
  "responses": {
...
    "SwaggerAllUserResponse": {
      "description": "UserResponse It is the response of the all users information",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
...
    "SwaggerErrorMessage": {
      "description": "ErrorMessage standardized error response.",
      "schema": {
        "$ref": "#/definitions/ErrorMessage"
      }
    },
...
    "SwaggerUserResponse": {
      "description": "UserResponse It is the response of the user's information",
      "schema": {
        "$ref": "#/definitions/UserResponse"
      }
    }
  }
}